[vfs] Make sticky bit check consistent with Linux.
Our implementation of vfs.CheckDeleteSticky was not consistent with Linux,
namely fs/linux.h:check_sticky(). One of the biggest differences was that vfs
implementation did not allow for the owner of the sticky directory to delete
any file inside it.

This change makes our implementation consistent with Linux.

Updates #3027

PiperOrigin-RevId: 354304402